+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to
 ## Unreleased
 #### Breaking Changes
 #### Added
-- rootfs/qemu.sh: support sata,nvme,virtio-blk.
+- rootfs/qemu.sh: support sata,nvme,virtio-blk,virtio-scsi.
 - wayland: Add helloworld.c demo.
 - python: add split.py,hexdump().
 - libvirt: Add pci xml demos.